fix: move blocking load diagnostics into blockingload subpackage for pkg-file-count
Extract blocking factory load errors and operator formatting into pkg/config/blockingload with a shared factoryerrors sentinel so pkg/config stays within the 15-file limit. Relocate materialization failure-baseline fixtures alongside the blockingload package.
