+++ pkg/config/factoryerrors/errors.go
@@ -1,0 +1,7 @@
+package factoryerrors
+
+import "errors"
+
+// ErrInvalidNamedFactory reports that the submitted named-factory payload could
+// not be normalized into a runnable named-factory layout.
+var ErrInvalidNamedFactory = errors.New("invalid named factory")
